--- src/Telas/Home/Componentes/BotaoDelivery.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { StyleSheet, TouchableOpacity } from 'react-native';
-
-export default function btnDelivery(){
-    return <TouchableOpacity>
-        
-    </TouchableOpacity>
-
-    
-}
