--- src/Telas/Home/Componentes/BotaoRetirar.js
@@ -1,18 +1,0 @@
-import React from "react";
-import { StyleSheet, TouchableOpacity, Text } from 'react-native';
-
-export default function btnRetira(){
-    return <TouchableOpacity>
-        
-    </TouchableOpacity>
-
-    
-}
-
-const estilos = StyleSheet.create({
-    botao:{
-
-    }
-
-
-})
