+++ src/Telas/Home/Components/BotaoDelivery.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { StyleSheet, TouchableOpacity } from 'react-native';
+
+export default function btnDelivery(){
+    return <TouchableOpacity>
+        
+    </TouchableOpacity>
+
+    
+}
