+++ src/Telas/Home/Components/BotaoRetirar.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { StyleSheet, TouchableOpacity, Text } from 'react-native';
+
+export default function btnRetira(){
+    return <TouchableOpacity>
+        
+    </TouchableOpacity>
+
+    
+}
+
+const estilos = StyleSheet.create({
+    botao:{
+
+    }
+
+
+})
