+++ src/Telas/telaCardapio/Cardapio.js
@@ -2,68 +2,63 @@ import React from "react";
 import { FlatList, StyleSheet, ScrollView } from "react-native";
 
 //Componentes padroes
-import Texto from "../../ComponentesPadroes/Texto";
+import Texto from "../../Components/Texto";
 
 // Componentes cardapio / botao, barra de pesquisa
-import BarraPesquisa from "./Componentes/Pesquisa";
-import Item from "./Componentes/itemPromo";
+import BarraPesquisa from "./Components/Pesquisa";
+import Item from "./Components/itemPromo";
 
+export default function Cardapio({ promocao, sabores }) {
+  const handlePesquisa = (termoPesquisa) => {
+    // Lógica de pesquisa aqui
+    console.log("Pesquisar por:", termoPesquisa);
+  };
 
-export default function Cardapio({promocao, sabores}){
-    const handlePesquisa = (termoPesquisa) => {
-        // Lógica de pesquisa aqui
-        console.log('Pesquisar por:', termoPesquisa);
-    };
-    
-    const renderItem = ({ item }) => {
-        return <Text>{item.nome}</Text>;
-    };
+  const renderItem = ({ item }) => {
+    return <Text>{item.nome}</Text>;
+  };
 
-    // return<>
-    // <BarraPesquisa onPesquisa={handlePesquisa} />
-    // <FlatList 
-    //     data={promocao.listaP}
-    //     renderItem={Item}
-    //     extractorKey={({nome})=>{nome}}
-            
-    //     />
-    // </>
+  // return<>
+  // <BarraPesquisa onPesquisa={handlePesquisa} />
+  // <FlatList
+  //     data={promocao.listaP}
+  //     renderItem={Item}
+  //     extractorKey={({nome})=>{nome}}
 
-    return<>
-        <FlatList 
-            data={promocao.listaP}
-            renderItem={Item}
-            extractorKey={({nome})=>{nome}}
-            ListHeaderComponent={()=>{
-                return <>
-                    <BarraPesquisa onPesquisa={handlePesquisa} />
-                </>
-            }}
-        />
+  //     />
+  // </>
+
+  return (
+    <>
+      <FlatList
+        data={promocao.listaP}
+        renderItem={Item}
+        extractorKey={({ nome }) => {
+          nome;
+        }}
+        ListHeaderComponent={() => {
+          return (
+            <>
+              <BarraPesquisa onPesquisa={handlePesquisa} />
+            </>
+          );
+        }}
+      />
     </>
+  );
 }
 
-
-
-
-
-
 const estilos = StyleSheet.create({
-    bemvindo:{
-        fontSize: 20,
-        color: "black",
-    },
-
-    
-})
-
-
-
+  bemvindo: {
+    fontSize: 20,
+    color: "black",
+  },
+});
 
 // export default function Cardapio({topo, detalhes, itens}){
 //     // return <ScrollView>
 //     //     <Topo {...topo}/>
-//     //     <Descricao {...detalhes}/> 
+//     //     <Descricao {...detalhes}/>
 //     //     <Itens {...itens} />
 //     // </ScrollView>
 //     return <>
@@ -81,4 +76,3 @@ const estilos = StyleSheet.create({
 //         />
 //     </>
 // }
-
